secao04/requirements.txt
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.25.0
bcrypt==3.2.2
black==22.3.0
cffi==1.15.1
click==8.1.3
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
uvicorn==0.17.6
